feat(types): add helper type for extracting nodes from IResolvers
