+++ package.json
@@ -46,6 +46,7 @@
     "husky": "^9.0.11",
     "lint-staged": "^15.2.0",
     "mercurius": "^16.1.0",
+    "mercurius-codegen": "^6.0.1",
     "pg": "^8.11.3",
     "prettier": "^3.1.1"
   },
